Move hello-world method; manage C++ name mangling of C interface.
